Fixed an boundary source face error.
When we have mixed reflecting/vacuum conditions, add_boundary_sources() was not correctly advancing the face index on the faces with vacuum conditions.  These are effectively phi_b = 0.0 faces, so no source should be added, but
if we want the capability to mix reflecting with surface sources later on, we need to correctly advance the face index on the purely vacuum faces.