Changing epetra/tpetra templating.
We now template solvers on the LinAlgType enum which
can be EPETRA or TPETRA.  This reduces the number of
template arguments from 2 (MV/OP) to 1 for most solvers,
and also simplifies the cases where a class needs to interact
with types related to the MV/OP types (e.g. maps or
specific matrix types).